push

+++ auth/models/usermodel.py
@@ -5,18 +5,16 @@ from auth.utils.utility import createInviteTemplate, createUserTemplate
 
 def usernameCheck(username):
    
-    for x in getUsersCollection().find({"username": f"{username}"}):
-        return True
-    return False
+    return bool(getUsersCollection().find_one({"username": f"{username}"}))
     
 
 def invCheck(invite):
-    for x in getInvitesCollection().find({"invite": f"{invite}"}):
-        return True
-    return False
-
+    
+    return bool(getInvitesCollection().find_one({"invite": f"{invite}"}))
+     
 
 def getInviter(invite):
+
     for x in getInvitesCollection().find({"invite": f"{invite}"}):
         inviter = x["inviter"]
         return inviter
@@ -24,7 +22,9 @@ def getInviter(invite):
 
 def login(username, password):
     if usernameCheck(username):
+
         for x in getUsersCollection().find({"username": f"{username}"}):
+
             hashedPassword = x["password"].encode("utf8")
             encodedPassword = password.encode("utf8")
 
@@ -34,7 +34,9 @@ def login(username, password):
             __isBanned__ = x["banned"]
             __inviter__ = x["invited_by"]
             __createdAt__ = x["created_at"]
+
             if bcrypt.checkpw(encodedPassword, hashedPassword):
+
                 ret = {
                     "status": "success",
                     "message": f"Successfully logged in as {__username__}",
@@ -63,6 +65,7 @@ def login(username, password):
             "message": "Invalid username"
         
         }
+
         return ret
 
 
@@ -72,17 +75,29 @@ def register(username, password, invite):
         inviter = getInviter(invite)
 
         if inviter:
-            encodedPassword = password.encode("utf8")
-            hashedPassword = bcrypt.hashpw(encodedPassword, bcrypt.gensalt())
-            getUsersCollection().insert_one(createUserTemplate(username, hashedPassword.decode(), inviter))
-            getInvitesCollection().delete_one({"invite": f"{invite}"})
 
-            ret = {
-                "status": "success",
-                "message": f"Successfully registered as {username}",
-            }
+            if usernameCheck(username) is False:
 
-            return ret
+                encodedPassword = password.encode("utf8")
+                hashedPassword = bcrypt.hashpw(encodedPassword, bcrypt.gensalt())
+
+                getUsersCollection().insert_one(createUserTemplate(username, hashedPassword.decode(), inviter))
+                getInvitesCollection().delete_one({"invite": f"{invite}"})
+
+                ret = {
+                    "status": "success",
+                    "message": f"Successfully registered as {username}",
+                }
+
+                return ret
+            else:
+
+                ret = {
+                    "status": "failed",
+                    "message": f"Username already taken",
+                }
+
+                return ret
     else:
 
         ret = {
@@ -91,3 +106,32 @@ def register(username, password, invite):
         }
 
         return ret
+
+def updatePassword(username, password, newPassword):
+
+    encodedPassword = newPassword.encode("ascii")
+    hashedPassword = bcrypt.hashpw(encodedPassword, bcrypt.gensalt())
+    __newPassword__ = hashedPassword.decode()
+
+    x = login(username, password)
+
+    if x["status"] == "success":
+
+        getUsersCollection().find_one_and_update({"username": f"{username}"},  { '$set': {"password": f"{__newPassword__}"} })
+
+        ret = {
+            "status": "success",
+            "message": f"Successfully changed the password of {username}",
+        }
+
+        return ret
+
+    else:
+
+        ret = {
+            "status": "failed",
+            "message": f"Invalid password",
+        }
+
+        return ret
+
